Moved player to entity package and removed character package
