+++ bullet_hell/src/main/java/pp2/GUI/GameFrame.java
@@ -6,7 +6,7 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Rectangle;
-import pp2.Character.Player;
+import pp2.Entity.Player;
 
 public class GameFrame {
     private final Rectangle gameFrame = new Rectangle();
