+++ bullet_hell/src/main/java/pp2/Resources/BackgroundImages.java
@@ -1,5 +1,5 @@
 package pp2.Resources;
 
 public class BackgroundImages {
-    private 
+
 }
